Update LICENSE (see commit description)
Context for this update: After completing this project, I discovered that there is a US Patent which covers some of the methods used in this video. If you are in the United States and are interested in a commercial license for the technology, you can reach out to the technology transfer office at University of Washington, CoMotion.